Extract streamv2 helper modules
